pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oamf"
version = "1.0.1.0"
authors = [
  { name="DEBELA", email="dabookoo@gmail.com" },
]
description = "oAMF package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "networkx",
  "pandas",
  "PyYAML"
]

[project.urls]
Homepage = "https://github.com/arg-tech/oamf"
Issues = "https://github.com/arg-tech/oamf/issues"